+++ CPanel/inputParams.h
@@ -44,7 +44,7 @@ struct inputParams
 
     using volMeshBounds_type = std::vector<double>;
     using volMeshBounds_index_type = volMeshBounds_type::size_type;
-    using volMeshRes_type = std::vector<int>;
+    using volMeshRes_type = std::vector<size_t>;
     using volMeshRes_index_type = volMeshRes_type::size_type;
 
     volMeshBounds_type volMeshBounds;
